Branching respects the sign of the relative address

+++ lib/instruction/bvc.c
@@ -1,7 +1,8 @@
+#include <stdint.h>
 #include "bvc.h"
 
 void bvc(struct cpu *cpu) {
 	if (!cpu->overflow) {
-		cpu->program_counter += cpu->operand;
+		cpu->program_counter += (int8_t)cpu->operand;
 	}
 }
